requirements_vanna.txt
vanna
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pandas==2.2.3
plotly==5.24.1
chromadb
google-generativeai==0.8.3
fastapi==0.115.6
uvicorn==0.34.0